move default app to new folder
