+++ angular.json
@@ -1,90 +1,86 @@
 {
   "$schema": "./node_modules/@angular/cli/lib/config/schema.json",
   "version": 1,
+  "newProjectRoot": "apps",
   "projects": {
     "my-app": {
-      "root": "",
-      "sourceRoot": "src",
+      "root": "apps/my-app",
+      "sourceRoot": "apps/my-app/src",
       "projectType": "application",
+      "schematics": {},
+      "prefix": "app",
       "architect": {
         "build": {
           "builder": "@angular/build:application",
           "options": {
-            "aot": false,
-            "allowedCommonJsDependencies": [
-              "highcharts",
-              "highcharts-custom-events"
-            ],
-            "outputPath": {"base": "dist"},
-            "index": "src/index.html",
-            "browser": "src/main.ts",
-            "tsConfig": "./tsconfig.app.json",
-            "polyfills": ["core-js/es/reflect", "zone.js"],
+            "outputPath": "dist/my-app",
+            "index": "apps/my-app/src/index.html",
+            "browser": "apps/my-app/src/main.ts",
+            "polyfills": ["zone.js"],
+            "tsConfig": "apps/my-app/tsconfig.app.json",
             "assets": [
-              "src/assets",
-              "src/favicon.ico"
+              {
+                "glob": "**/*",
+                "input": "apps/my-app/public"
+              }
             ],
-            "styles": [],
-            "scripts": [],
-            "extractLicenses": false,
-            "sourceMap": true,
-            "optimization": false,
-            "namedChunks": true
+            "styles": ["apps/my-app/src/styles.css"],
+            "scripts": []
           },
           "configurations": {
             "production": {
               "budgets": [
                 {
+                  "type": "initial",
+                  "maximumWarning": "500kB",
+                  "maximumError": "1MB"
+                },
+                {
                   "type": "anyComponentStyle",
-                  "maximumWarning": "6kb"
+                  "maximumWarning": "4kB",
+                  "maximumError": "8kB"
                 }
               ],
-              "optimization": true,
-              "outputHashing": "all",
-              "sourceMap": false,
-              "namedChunks": false,
-              "aot": true,
-              "extractLicenses": true,
-              "fileReplacements": [
-                {
-                  "replace": "src/environments/environment.ts",
-                  "with": "src/environments/environment.prod.ts"
-                }
-              ]
+              "outputHashing": "all"
+            },
+            "development": {
+              "optimization": false,
+              "extractLicenses": false,
+              "sourceMap": true
             }
           },
-          "defaultConfiguration": ""
+          "defaultConfiguration": "production"
         },
         "serve": {
           "builder": "@angular/build:dev-server",
-          "options": {
-            "buildTarget": "my-app:build"
-          },
           "configurations": {
             "production": {
               "buildTarget": "my-app:build:production"
+            },
+            "development": {
+              "buildTarget": "my-app:build:development"
             }
-          }
+          },
+          "defaultConfiguration": "development"
         },
         "extract-i18n": {
-          "builder": "@angular/build:extract-i18n",
-          "options": {
-            "buildTarget": "my-app:build"
-          }
+          "builder": "@angular/build:extract-i18n"
         },
         "test": {
           "builder": "@angular-devkit/build-angular:karma",
           "options": {
-            "main": "src/test.ts",
+            "main": "apps/my-app/src/test.ts",
             "karmaConfig": "./karma.conf.js",
-            "polyfills": ["core-js/es/reflect", "zone.js"],
-            "tsConfig": "./tsconfig.spec.json",
-            "scripts": [],
-            "styles": [],
+            "polyfills": ["core-js/es/reflect", "zone.js", "zone.js/testing"],
+            "tsConfig": "apps/my-app/tsconfig.spec.json",
             "assets": [
-              "src/assets",
-              "src/favicon.ico"
-            ]
+              {
+                "glob": "**/*",
+                "input": "apps/my-app/public"
+              }
+            ],
+            "styles": ["apps/my-app/src/styles.css"],
+            "scripts": []
           }
         }
       }
