+++ apps/my-app/src/app/app.config.ts
@@ -1,0 +1,28 @@
+import {provideHttpClient, withInterceptorsFromDi} from '@angular/common/http';
+import {ApplicationConfig, importProvidersFrom, provideZoneChangeDetection} from '@angular/core';
+import {BrowserModule} from '@angular/platform-browser';
+import { provideHighCharts } from 'highcharts-angular';
+import ExportingModule from 'highcharts/modules/exporting';
+import SunsetTheme from 'highcharts/themes/sunset';
+
+export const appConfig: ApplicationConfig = {
+  providers: [
+    provideZoneChangeDetection({ eventCoalescing: true }),
+    importProvidersFrom(BrowserModule),
+    provideHttpClient(withInterceptorsFromDi()),
+    provideHighCharts({
+      options: {
+        title: {
+          style: {
+            color: 'tomato'
+          }
+        },
+        legend: {
+          enabled: false
+        }
+      },
+      // The modules will work for all charts.
+      modules: [ExportingModule, SunsetTheme]
+    })
+  ]
+};
