+++ apps/my-app/src/app/apple-data.service.ts
