+++ apps/my-app/src/app/gantt-chart/gantt-chart.component.css
