+++ apps/my-app/src/app/lazy-loading-chart/lazy-loading-chart.component.css
