+++ apps/my-app/src/app/stock-chart/stock-chart.component.ts
