--- tsconfig.app.json
@@ -1,19 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../out-tsc/app",
-    "baseUrl": "tsconfig.app",
-    "types": [],
-    "paths": {
-      "highcharts-angular": [
-        "../dist/highcharts-angular"
-      ],
-      "highcharts-angular/*": [
-        "../dist/highcharts-angular/*"
-      ]
-    }
-  },
-  "files": [
-    "src/main.ts"
-  ]
-}
