--- tsconfig.spec.json
@@ -1,23 +1,0 @@
-/* To learn more about this file see: https://angular.io/config/tsconfig. */
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "paths": {
-      "highcharts-angular": ["../dist/highcharts-angular", "highcharts-angular/src/public_api.ts"],
-      "highcharts-angular/*": ["../dist/highcharts-angular/*"]
-    },
-    "outDir": "./out-tsc/spec",
-    "types": [
-      "jasmine",
-      "node"
-    ]
-  },
-  "files": [
-    "src/test.ts"
-  ],
-  "include": [
-    "src/**/*.spec.ts",
-    "src/**/*.d.ts"
-  ]
-}
-
